pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vidore_benchmark"]

[project]
name = "vidore_benchmark"
version = "0.0.10"
description = "Visual Document Retrieval: Benchmark"
authors = [
    { name = "Manuel Faysse", email = "manuel.faysse@illuin.tech" },
    { name = "Hugues Sibille", email = "hugues.sibille@illuin.tech" },
    { name = "Tony Wu", email = "tony.wu@illuin.tech" },
]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "datasets>=2.15.0,<3.0.0",
    "einops>=0.8.0,<1.0.0",
    "GPUtil>=1.4.0,<2.0.0",
    "mteb>=1.12.47,<2.0.0",
    "numpy>=1.21.2,<2.0.0",
    "pdf2image>=1.17.0,<2.0.0",
    "peft>=0.11.1,<1.0.0",
    "pillow>=9.2.0,<11.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "sentencepiece>=0.2.0,<1.0.0",
    "torch>=2.0.0,<3.0.0",
    "transformers>=4.41.1,<5.0.0",
    "typer>=0.12.3,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.4.2",
    "ipykernel>=6.29.4",
    "mypy>=1.10.0",
    "pytest>=8.2.1",
    "ruff>=0.4.5",
]

bge-m3 = ["FlagEmbedding>=1.2.10,<2.0.0"]
bm25 = ["nltk>=3.8.1,<4.0.0", "rank-bm25>=0.2.2,<1.0.0"]
jina-clip = ["timm>=1.0.7,<2.0.0"]
sentence-transformers = ["sentence-transformers>=3.0.1,<4.0.0"]
all-retrievers = [
    "rank-bm25>=0.2.2,<1.0.0",
    "FlagEmbedding>=1.2.10,<2.0.0",
    "timm>=1.0.7,<2.0.0",
    "nltk>=3.8.1,<4.0.0",
    "sentence-transformers>=3.0.1,<4.0.0",
]

interpretability = ["matplotlib>=3.9.0,<4.0.0", "seaborn>=0.13.2,<1.0.0"]

baselines = [
    "accelerate==0.30.1",
    "datasets==2.19.1",
    "einops==0.8.0",
    "FlagEmbedding==1.2.10",
    "GPUtil==1.4.0",
    "mteb>=1.12.47",
    "nltk==3.8.1",
    "numpy==1.26.4",
    "pdf2image==1.17.0",
    "peft==0.11.1",
    "pillow==10.3.0",
    "pytesseract==0.3.10",
    "python-dotenv==1.0.1",
    "rank-bm25==0.2.2",
    "sentence-transformers==3.0.1",
    "sentencepiece==0.2.0",
    "torch==2.2.2",
    "transformers==4.41.1",
]

[project.urls]
homepage = "https://github.com/tonywu71/vidore-benchmark"

[project.scripts]
vidore-benchmark = "vidore_benchmark.main:app"
generate-similarity-maps = "vidore_benchmark.generate_similarity_maps:app"

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120
